requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
requests==2.31.0
websocket-client==1.6.4
SQLAlchemy==1.4.53
Werkzeug==2.3.7
python-dotenv==1.0.0
cryptography==41.0.7
PyJWT==2.8.0
gunicorn==21.2.0
stripe==7.8.0
